Refactored:   Componentized `<GenreCard />`'s player mechanisms (src/components/genres/GenreCard.tsx) Added:    A component to encapsulate player logics and animations (src/components/genres/GenrePlayer.tsx)
